moved selectUser from app to joinlist
